tests/wpt/tests/fedcm/fedcm-identity-assertion-nocors.https.html

<!DOCTYPE html>
<title>Federated Credential Management API test with no CORS identity assertion.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
import {request_options_with_mediation_required,
        fedcm_test,
        select_manifest,
        mark_signed_in,
        fedcm_get_dialog_type_promise,
        fedcm_get_and_select_first_account} from './support/fedcm-helper.sub.js';

fedcm_test(async t => {
  await mark_signed_in();
  let test_options = request_options_with_mediation_required("manifest-token-nocors.json");
  await select_manifest(t, test_options);
  try {
    const cred = await fedcm_get_and_select_first_account(t, test_options);
    assert_unreached("An IdentityCredentialError exception should be thrown.");
  } catch (e) {
    assert_true(e instanceof DOMException);
    assert_equals(e.name, "IdentityCredentialError");
  }
}, 'Test that promise is rejected if identity assertion does not use CORS');
</script>
